fixed proejct search results
